orbito-erp/backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
supabase>=2.8,<3.0
python-dotenv>=1.0,<2.0
passlib[bcrypt]>=1.7,<2.0
bcrypt>=4.0,<4.1
python-jose[cryptography]>=3.3,<4.0
httpx>=0.27,<1.0
google-genai>=1.0,<2.0
pypdf>=5.0,<6.0
python-multipart>=0.0.9,<1.0
pydantic[email]>=2.9,<3.0
pytest>=8.3,<9.0
